requirements.txt
alembic==1.8.1
aniso8601==9.0.1  
autopep8==1.7.0   
click==8.1.3      
colorama==0.4.5   
Flask==2.2.2      
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.0
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.4
pycodestyle==2.9.1
PyJWT==2.5.0
Pillow==9.2.0
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
SQLAlchemy==1.4.41
toml==0.10.2
Werkzeug==2.2.2
gunicorn==20.0.4